server/requirements.txt
# Основные библиотеки
Flask==2.2.3
Flask-RESTful==0.3.9
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Werkzeug==2.2.3
gunicorn==20.1.0

# Базы данных
sqlalchemy==1.4.46
psycopg2-binary==2.9.5
alembic==1.9.2

# Безопасность
passlib==1.7.4
bcrypt==4.0.1
pyjwt==2.6.0

# Discord интеграция
discord.py==2.1.0

# Утилиты
python-dotenv==0.21.1
pydantic==1.10.5
requests==2.28.2
schedule==1.1.0 